# Chore: Resolves #12088: Desktop: Add performance logging statements to the startup code