# Summary

This pull request refactors the desktop application startup logic to log the time each startup task takes.

Follow-up to https://github.com/laurent22/joplin/pull/12776.
Resolves #12088.

# Testing plan

1. Start the desktop application.
2. Open the development tools.
3. Open the "Console" tab.
4. Filter by "Performance".
5. Verify that log output similar to the following is visible:
    <details><summary>Output</summary>
    
    ```
    13:44:13: Performance: main-html/main: Start at 1s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/start: Start at 1.01s
    Logger.ts:317 13:44:13: Performance: BaseApplication/start: Start at 1.01s
    Logger.ts:317 13:44:13: Performance: BaseApplication/start: End at 1.05s (took 40.5ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up extra debug logging: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up extra debug logging: End at 1.06s (took 0.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up registry: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up registry: End at 1.06s (took 0.1ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up auto updater: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up auto updater: End at 1.06s (took 0.1ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up AlarmService: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up AlarmService: End at 1.06s (took 0.1ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/openDevTools: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/openDevTools: End at 1.06s (took 0.5ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up custom protocol handler: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up custom protocol handler: End at 1.06s (took 3.4ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initialize PluginManager, redux, CommandService, and KeymapService: Start at 1.06s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initialize PluginManager, redux, CommandService, and KeymapService: End at 1.07s (took 7.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initialize PerFolderSortOrderService: Start at 1.07s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initialize PerFolderSortOrderService: End at 1.07s (took 0.8ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/dispatch initial settings: Start at 1.07s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/dispatch initial settings: End at 1.08s (took 6.6ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/update folders and tags: Start at 1.08s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/update folders and tags: End at 1.09s (took 9.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up custom CSS: Start at 1.09s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up custom CSS: End at 1.09s (took 0.5ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/send initial selection to redux: Start at 1.09s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/send initial selection to redux: End at 1.09s (took 1.2ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initializeUserFetcher: Start at 1.09s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initializeUserFetcher: End at 1.09s (took 0.2ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/updateTray: Start at 1.09s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/updateTray: End at 1.09s (took 1.6ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set main window state: Start at 1.09s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set main window state: End at 1.1s (took 5.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/start maintenance tasks: Start at 1.1s
    Logger.ts:317 13:44:13: Performance: ShareService/maintenance: Start at 1.1s
    Logger.ts:317 13:44:13: Performance: ShareService/maintenance: End at 1.1s (took 0.4ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/start maintenance tasks: End at 1.1s (took 2.4ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up ClipperServer: Start at 1.1s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up ClipperServer: End at 1.11s (took 8.1ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up external edit watchers: Start at 1.11s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up external edit watchers: End at 1.11s (took 0.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/add debug variables: Start at 1.11s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/add debug variables: End at 1.11s (took 0.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/listen for main process events: Start at 1.11s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/listen for main process events: End at 1.11s (took 2ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initPluginService: Start at 1.11s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/initPluginService: End at 1.26s (took 151.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/setupContextMenu: Start at 1.26s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/setupContextMenu: End at 1.29s (took 27.8ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up SpellCheckerService: Start at 1.29s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/set up SpellCheckerService: End at 1.31s (took 17.1ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/listen for resource events: Start at 1.31s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/listen for resource events: End at 1.31s (took 0.5ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/setupOcrService: Start at 1.31s
    Logger.ts:317 13:44:13: Performance: app-desktop/app/setupOcrService: End at 1.31s (took 0.3ms)
    Logger.ts:317 13:44:13: Performance: app-desktop/app/start: End at 1.31s (took 0.3s)
    Logger.ts:317 13:44:13: Performance: main-html/main: End at 1.53s (took [ 0.53s ])
    Logger.ts:317 13:44:14: Performance: DecryptionWorker/start: Start at 2.11s
    Logger.ts:317 13:44:14: Performance: DecryptionWorker/start: End at 2.11s (took 0.8ms)
    ```
    </details>
6. Navigate to the "Performance" tab of the development tools.
7. Click "Record and reload".
8. Verify that the "Timings" group shows tasks from the main startup logic:
  <img width="2880" height="470" alt="screenshot: Timings section includes tasks with names starting with main-html/, app-desktop/app/, etc. These tasks all start after about 1.1s (do not include the first 1.5s of page load time)" src="https://github.com/user-attachments/assets/586bb583-4cf1-4301-acc2-e02d6c39bd60" />

  - **Note**: The tasks listed in the chart all start after about 1.1s. Content that runs before this includes `require`s from the top of imported files.

<!--

Please prefix the title with the platform you are targetting:

Here are some examples of good titles:

- Desktop: Resolves #123: Added new setting to change font
- Mobile, Desktop: Fixes #456: Fixed config screen error
- All: Resolves #777: Made synchronisation faster

And here's an explanation of the title format:

- "Desktop" for the Windows/macOS/Linux app (Electron app)
- "Mobile" for the mobile app (or "Android" / "iOS" if the pull request only applies to one of the mobile platforms)
- "CLI" for the CLI app

If it's two platforms, separate them with commas - "Desktop, Mobile" or if it's for all platforms, prefix with "All".

If it's not related to any platform (such as a translation, change to the documentation, etc.), simply don't add a platform.

Then please append the issue that you've addressed or fixed. Use "Resolves #123" for new features or improvements and "Fixes #123" for bug fixes.

AND PLEASE READ THE GUIDE: https://github.com/laurent22/joplin/blob/dev/readme/dev/index.md

-->